+++ libraries/api-client/src/client/clientError.ts
