+++ libraries/api-client/src/client/publicClient.ts
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2021 Wire Swiss GmbH
+ * Copyright (C) 2018 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,8 +17,9 @@
  *
  */
 
-import {ClientCapability} from './ClientCapability';
+import {ClientClassification} from './clientClassification';
 
-export type ClientCapabilityData = {
-  capabilities?: ClientCapability[];
-};
+export interface PublicClient {
+  class: ClientClassification;
+  id: string;
+}
