+++ libraries/api-client/src/client/registeredClient.ts
